+++ SasquatchMac/SasquatchMac/AppCenterDelegate.swift
@@ -39,6 +39,8 @@ import AppCenter
   func trackPage(_ pageName: String, withProperties: Dictionary<String, String>)
   func resume()
   func pause()
+  func startSession()
+  func enableManualSessionTracker()
 
   // Crashes section.
   func hasCrashedInLastSession() -> Bool