+++ SasquatchMac/SasquatchMacObjC/AppCenterDelegateObjC.m
@@ -120,6 +120,14 @@
   [MSACAnalytics pause];
 }
 
+- (void)startSession {
+  [MSACAnalytics startSession];
+}
+
+- (void)enableManualSessionTracker {
+  [MSACAnalytics enableManualSessionTracker];
+}
+
 #pragma mark - MSACCrashes section.
 
 - (BOOL)hasCrashedInLastSession {