+++ SasquatchMac/SasquatchMacSwift/AppDelegate.swift
@@ -69,6 +69,11 @@ class AppDelegate: NSObject, NSApplicationDelegate, CrashesDelegate, CLLocationM
       AppCenter.logUrl = logUrl
     }
     
+    // Set manual session tracker before App Center start.
+    if UserDefaults.standard.bool(forKey: kMSManualSessionTracker) {
+      Analytics.enableManualSessionTracker()
+    }
+    
     // Set location manager.
     locationManager.delegate = self
     locationManager.desiredAccuracy = kCLLocationAccuracyKilometer